test(SearchInput): move tests to own file
